requirements.txt
djangorestframework
Pillow
python-decouple
drf-yasg
psycopg2-binary
python-slugify

